+++ app/src/main/kotlin/net/canvoki/vokibot/MainActivity.kt
@@ -3,8 +3,6 @@ package net.canvoki.vokibot
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.foundation.layout.padding
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import net.canvoki.shared.component.AppScaffold
 import net.canvoki.shared.component.StackNavigator